Uploaded jfm_dupeRig_parentConstrain
Maya Python script opens a GUI window, allows user to select and duplicate a joint rig then parent each original joint to the corresponding new joint.  A control for all constraints will be added to the root of the constrained rig, the new rig will be cleaned of all shapes, transforms and effectors.