Parents an AUX effector under a control shape
Script to parent an HIK AUX effector under a control shape, then links AUX effector translation reach and rotation reach to two attributes, one on the control shape and another on a master shape so that you can turn the effect of the AUX handle on and off either locally or globally.  Also hides the AUX effector based on an attribute on the master control shape.